aggregation items to ArrayImagePlot
