+++ src/silx/gui/data/NXdataWidgets.py
@@ -35,6 +35,7 @@ from silx.gui.data.NumpyAxesSelector import NumpyAxesSelector
 from silx.gui.plot import Plot1D, Plot2D, StackView, ScatterView, items
 from silx.gui.plot.ComplexImageView import ComplexImageView
 from silx.gui.plot.items.image_aggregated import ImageDataAggregated
+from silx.gui.plot.actions.image import AggregationModeAction
 from silx.gui.colors import Colormap
 from silx.gui.widgets.FrameBrowser import HorizontalSliderWithBrowser
 
@@ -423,7 +424,25 @@ class ArrayImagePlot(qt.QWidget):
         layout.addWidget(self._auxSigSlider)
 
         self.setLayout(layout)
+        
+        self.__aggregationModeAction = AggregationModeAction(parent=self)
+        self.getPlot().toolBar().addAction(self.__aggregationModeAction)
+        self.__aggregationModeAction.sigAggregationModeChanged.connect(self._aggregationModeChanged)
 
+    def getAggregationModeAction(self) -> AggregationModeAction:
+        """Action toggling the aggregation mode action
+        """
+        return self.__aggregationModeAction
+
+    def _aggregationModeChanged(self):
+        item = self.getPlot()._getItem("image")
+
+        if item is None:
+            return
+        
+        if isinstance(item, ImageDataAggregated):
+            item.setAggregationMode(self.getAggregationModeAction().getAggregationMode())
+            
     def _sliderIdxChanged(self, value):
         self._updateImage()
 
@@ -556,10 +575,7 @@ class ArrayImagePlot(qt.QWidget):
                 "image",
             )
         )
-        imageItem = ImageDataAggregated()
-        imageItem.setName(legend)
-        self._plot.addItem(imageItem)
-        
+
         if xcalib.is_affine() and ycalib.is_affine():
             # regular image
             xorigin, xscale = xcalib(0), xcalib.get_slope()
@@ -570,11 +586,14 @@ class ArrayImagePlot(qt.QWidget):
             self._plot.getXAxis().setScale("linear")
             self._plot.getYAxis().setScale("linear")
             
+            imageItem = ImageDataAggregated()
+            imageItem.setName(legend)
             imageItem.setData(image)
             imageItem.setOrigin(origin)
             imageItem.setScale(scale)
             imageItem.setColormap(self._plot.getDefaultColormap())
-
+            imageItem.setAggregationMode(self.getAggregationModeAction().getAggregationMode())
+            self._plot.addItem(imageItem)
         else:
             xaxisscale, yaxisscale = self._axis_scales
